model-training/ball-detection-improvement/yolo/modules/trainer.py
"""Training module that wraps Ultralytics YOLO with reporting helpers."""

from __future__ import annotations

import logging
from pathlib import Path
from typing import Any, Dict, Optional

import yaml
from ultralytics import YOLO

logger = logging.getLogger(__name__)


class YOLOTrainer:
    """Handles YOLO model training, evaluation, and reporting."""

    def __init__(
        self,
        yolo_config: Dict[str, Any],
        paths_config: Dict[str, Any],
        device: str,
        amp_enabled: bool = True,
    ):
        self.yolo_config = yolo_config.copy()
        self.paths_config = paths_config
        self.device = device
        self.amp_enabled = amp_enabled
        self.model: Optional[YOLO] = None

    # ------------------------------------------------------------------
    def train(self, dataset_yaml_path: Path) -> Optional[Path]:
        logger.info("🏋️ Starting YOLO training ...")

        training_config = self._prepare_training_config(dataset_yaml_path)
        self._log_training_config(training_config)

        model_arch = training_config.pop("model")
        self.model = YOLO(model_arch)

        baseline_results = self._evaluate_baseline(dataset_yaml_path)

        try:
            results = self.model.train(**training_config)
            logger.info("✅ Training completed successfully")

            models_root = self._get_models_root()
            project_name = self.yolo_config.get("project_name", "player_ball_detector")
            project_dir = models_root / project_name
            best_model_path = project_dir / "weights" / "best.pt"

            try:
                self._generate_training_reports(
                    results, project_dir, best_model_path, baseline_results
                )
            except Exception as report_error:  # pragma: no cover - reporting optional
                logger.warning(f"Failed to generate reports: {report_error}")

            if best_model_path.exists():
                return best_model_path

            logger.warning("Best model was not found at expected path; check training logs")
            return None

        except Exception as exc:  # pragma: no cover - training errors bubble up
            logger.error(f"❌ Training failed: {exc}")
            raise

    def evaluate(self, model_path: Optional[Path] = None) -> Dict[str, Any]:
        model = self._resolve_model(model_path)
        logger.info("📊 Evaluating model ...")
        return model.val()

    def predict(
        self,
        source: str,
        model_path: Optional[Path] = None,
        save: bool = True,
        conf: Optional[float] = None,
    ) -> Any:
        model = self._resolve_model(model_path)
        logger.info(f"🔍 Running inference on {source}")
        predict_args = {
            "source": source,
            "save": save,
            "device": self.device,
            "conf": conf or self.yolo_config.get("conf", 0.01),
        }
        return model.predict(**predict_args)

    # ------------------------------------------------------------------
    def _prepare_training_config(self, dataset_yaml_path: Path) -> Dict[str, Any]:
        config = self.yolo_config.copy()
        config["data"] = str(dataset_yaml_path)
        config["device"] = self.device
        config["amp"] = self.amp_enabled

        models_root = self._get_models_root()
        config["project"] = str(models_root)
        project_name = config.pop("project_name", "player_ball_detector")
        config["name"] = project_name
        return config

    def _log_training_config(self, config: Dict[str, Any]) -> None:
        logger.info("Training configuration summary:")
        logger.info("-" * 60)
        for key in sorted(config.keys()):
            if key in {"data", "project"}:
                continue
            logger.info(f"{key:20s} : {config[key]}")
        logger.info("-" * 60)

    def _get_models_root(self) -> Path:
        workspace_root = Path(self.paths_config["workspace_root"]).expanduser()
        return workspace_root / self.paths_config.get("models_root", "models")

    def _get_reports_root(self) -> Path:
        workspace_root = Path(self.paths_config["workspace_root"]).expanduser()
        return workspace_root / self.paths_config.get("reports_root", "reports")

    def _resolve_model(self, model_path: Optional[Path]) -> YOLO:
        if model_path and Path(model_path).exists():
            return YOLO(str(model_path))
        if self.model:
            return self.model
        raise ValueError("No model is available; train or provide weights path first")

    def _generate_training_reports(
        self,
        results: Any,
        project_dir: Path,
        model_path: Optional[Path],
        baseline_results: Optional[Dict[str, Any]] = None,
    ) -> None:
        try:
            from reports.report_generator import EnhancedReportGenerator

            reports_root = self._get_reports_root()
            project_name = self.yolo_config.get("project_name", "player_ball_detector")
            generator = EnhancedReportGenerator(str(reports_root), project_name)

            dataset_stats = self._collect_dataset_stats()
            final_metrics = self._extract_final_metrics(results, model_path)
            training_data = {
                "config": {
                    "yolo_params": self.yolo_config,
                    "paths": self.paths_config,
                    "device": self.device,
                    "amp_enabled": self.amp_enabled,
                },
                "dataset_stats": dataset_stats,
                "model_path": str(model_path) if model_path else None,
            }

            generator.generate_training_reports(
                training_data=training_data,
                baseline_metrics=baseline_results,
                final_metrics=final_metrics,
            )

            reports_dir = project_dir / "reports"
            reports_dir.mkdir(exist_ok=True)
            excel_report = reports_root / project_name
            if excel_report.exists():
                for report_file in excel_report.rglob("training_report_*.html"):
                    target = reports_dir / report_file.name
                    if target.exists():
                        target.unlink()
                    target.symlink_to(report_file)
        except ImportError:
            logger.info("Report generation skipped (dependencies missing)")

    def _collect_dataset_stats(self) -> Dict[str, Any]:
        stats: Dict[str, Any] = {}
        dataset_yaml = self.yolo_config.get("data")
        if not dataset_yaml or not Path(dataset_yaml).exists():
            return stats

        with open(dataset_yaml, "r", encoding="utf-8") as handle:
            data = yaml.safe_load(handle)

        dataset_root = Path(data["path"])
        for split in ["train", "val", "test"]:
            images_dir = dataset_root / split / "images"
            if images_dir.exists():
                stats[f"{split}_images"] = len(list(images_dir.glob("*.jpg")))
        stats["class_names"] = data.get("names", [])
        stats["nc"] = data.get("nc", len(stats["class_names"]))
        return stats

    def _evaluate_baseline(self, dataset_yaml_path: Path) -> Dict[str, Any]:
        baseline_metrics = {
            "baseline_map50": 0.0,
            "baseline_player_map50": 0.9,
            "baseline_player_precision": 0.95,
            "baseline_player_recall": 0.9,
            "baseline_ball_map50": 0.0,
        }
        try:
            if not self.model:
                self.model = YOLO(self.yolo_config["model"])
            results = self.model.val(
                data=str(dataset_yaml_path),
                split="val",
                save=False,
                verbose=False,
            )
            if hasattr(results, "box") and hasattr(results.box, "map50"):
                baseline_metrics["baseline_map50"] = float(results.box.map50)
        except Exception as exc:
            logger.debug("Baseline evaluation failed: %s", exc)
        return baseline_metrics

    def _extract_final_metrics(self, results: Any, model_path: Optional[Path]) -> Dict[str, Any]:
        metrics: Dict[str, Any] = {}
        try:
            if model_path and Path(model_path).exists():
                final_model = YOLO(str(model_path))
                dataset_yaml = self.yolo_config.get("data")
                if dataset_yaml:
                    val_results = final_model.val(data=dataset_yaml, save=False)
                    if hasattr(val_results, "box"):
                        metrics["map50"] = float(getattr(val_results.box, "map50", 0.0))
                        metrics["precision"] = float(getattr(val_results.box, "mp", 0.0))
                        metrics["recall"] = float(getattr(val_results.box, "mr", 0.0))
        except Exception as exc:
            logger.debug("Final metrics extraction failed: %s", exc)

        if not metrics and hasattr(results, "results_dict"):
            metrics = results.results_dict
        return metrics


def create_trainer(
    yolo_config: Dict[str, Any],
    paths_config: Dict[str, Any],
    device: str,
    amp_enabled: bool = True,
) -> YOLOTrainer:
    return YOLOTrainer(yolo_config, paths_config, device, amp_enabled)
