model-training/ball-detection-improvement/requirements.txt
PyYAML>=6.0
Pillow>=10.0
tqdm>=4.66
